+++ res/win32_clipboard_text_html.ps1
@@ -1,0 +1,5 @@
+Add-Type -Assembly PresentationCore
+$content = [Windows.Clipboard]::GetData("HTML Format")
+
+[Console]::OutputEncoding = [System.Text.Encoding]::UTF8
+[Console]::WriteLine($content)